Split out array typing declarations
